Added: Start of the UI system. 'Windows' now appear when hovering over a tower and can be pinned via clicking the hovered tower (windows then can only be closed by clicking the close button on them)
TextMeshPro. Unity's default text is horrible and unreadable at such a lower canvas resolution (for the tower windows) so we have to use this for any UI text element.

Names for each tower defined in Tower.cs and set by the inheriting script. This is so the window can display the name of the tower on it's header.

Prefab folder to store prefabs (Resources/Prefabs)

Changes:
Segregated the code for hover detection in Mouse.cs to it's own function (stops Update function from becoming bloated and unmaintainable). Also did the same for the new clicking function (added with basic UI system)